third_party/blink/web_tests/wpt_internal/private-aggregation/shared-storage-context-id-sends-report.https.html

<!doctype html>
<meta name=timeout content=long>
<script src="/common/get-host-info.sub.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<script src="/shared-storage/resources/util.js"></script>

<body>
<script>
'use strict';

private_aggregation_promise_test(async () => {
  await addModuleOnce("resources/shared-storage-module.js");

  const data = {enableDebugMode: true, contributions: [{bucket: 1n, value: 2}]};
  await sharedStorage.run(
      "contribute-to-histogram",
      {data, keepAlive: true,
       privateAggregationConfig: {contextId: "example-context-id"}});

  const reports = await pollReports(
      "/.well-known/private-aggregation/report-shared-storage",
      /*wait_for=*/1, /*timeout=*/6000)
  assert_equals(reports.length, 1);

  const report = JSON.parse(reports[0]);
  verifyReport(report, /*api=*/'shared-storage', /*is_debug_enabled=*/true,
               /*debug_key=*/undefined,
               /*expected_cleartext_payload=*/ONE_CONTRIBUTION_EXAMPLE_PAYLOAD,
               /*expected_context_id=*/"example-context-id");

  const debug_reports = await pollReports(
      "/.well-known/private-aggregation/debug/report-shared-storage")
  assert_equals(debug_reports.length, 1);

  verifyReportsIdenticalExceptPayload(report, JSON.parse(debug_reports[0]));

}, 'run() that calls Private Aggregation with context ID');

private_aggregation_promise_test(async () => {
  await addModuleOnce("resources/shared-storage-module.js");

  const data = {enableDebugMode: false, contributions: [{bucket: 1n, value: 2}]};
  await sharedStorage.run(
      "contribute-to-histogram",
      {data, keepAlive: true,
       privateAggregationConfig: {contextId: "example-context-id"}});

  const reports = await pollReports(
      "/.well-known/private-aggregation/report-shared-storage",
      /*wait_for=*/1, /*timeout=*/6000)
  assert_equals(reports.length, 1);

  const report = JSON.parse(reports[0]);
  verifyReport(report, /*api=*/'shared-storage', /*is_debug_enabled=*/false,
               /*debug_key=*/undefined,
               /*expected_cleartext_payload=*/undefined,
               /*expected_context_id=*/"example-context-id");

  // We use a short timeout as the previous poll should've waited long enough.
  const debug_reports = await pollReports(
      "/.well-known/private-aggregation/debug/report-shared-storage",
      /*wait_for=*/1, /*timeout=*/50)
  assert_equals(debug_reports, null);
}, 'run() that calls Private Aggregation with context ID and no debug mode');

private_aggregation_promise_test(async () => {
  await addModuleOnce("resources/shared-storage-module.js");

  const data = {enableDebugMode: false, contributions: []};
  await sharedStorage.run(
      "contribute-to-histogram",
      {data, keepAlive: true,
       privateAggregationConfig: {contextId: "example-context-id"}});

  const reports = await pollReports(
      "/.well-known/private-aggregation/report-shared-storage",
      /*wait_for=*/1, /*timeout=*/6000)
  assert_equals(reports.length, 1);

  const report = JSON.parse(reports[0]);
  verifyReport(report, /*api=*/'shared-storage', /*is_debug_enabled=*/false,
               /*debug_key=*/undefined,
               /*expected_cleartext_payload=*/undefined,
               /*expected_context_id=*/"example-context-id");

  // We use a short timeout as the previous poll should've waited long enough.
  const debug_reports = await pollReports(
      "/.well-known/private-aggregation/debug/report-shared-storage",
      /*wait_for=*/1, /*timeout=*/50)
  assert_equals(debug_reports, null);
}, 'run() that calls Private Aggregation with context ID and no contributions');

</script>
</body>
